Update form component imports and remove unused code
